requirements.txt
fastapi
uvicorn[standard]
arxiv
python-dateutil
openai
httpx
python-dotenv